♻️ WOW-56 refactor : zustand 선언 파일 구조 변경
